+++ dcPower/main.c
@@ -21,7 +21,7 @@ extern interrupt void adcIsr(void);
 extern interrupt void sciaRxFifoIsr(void);
 extern interrupt void sciaTxFifoIsr(void);
 
-float Vdc_fnd_data;
+double Vdc_fnd_data;
 
 void main( void )
 {
@@ -29,7 +29,7 @@ void main( void )
     // int i;
     int trip_code,loop_ctrl,temp,count;
 	int cmd;
-	float ref_in0;
+	double ref_in0;
 
 	InitSysCtrl();
 	InitGpio();
